cloud-functions/requirements.txt
google.cloud
google-api-core
google-cloud-storage
google-cloud-datastore==2.0.1
google-api-python-client
google-cloud-pubsub
google-cloud-runtimeconfig
google-cloud-logging
pyyaml
pymysql
cryptography==3.2
python-jose[cryptography]==3.0.1
oauth2client
netaddr